Track all slash commands [#94]
